requirements.txt
behave==1.2.6
parse==1.11.1
parse-type==0.4.2
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
PyHamcrest==2.0.2
allure-behave==2.8.34
allure-python-commons==2.8.34
appdirs==1.4.3
Appium-Python-Client==0.36
requests-futures==1.0.0
image==1.5.33
pytesseract==0.3.7
requests==2.25.1
nose==1.3.7